Show symlinked files and dirs correctly (#72)
* Treat symlinks as normal files and dirs

* Dirs are recognized as dirs no matter is they are a symlink
  (correct devicon)
* Permission flags shown in the footer are the permissions of the target
  in case a symlink is selected
* Size of a symlinked file is the size of the target
* File display to be fixed: symlink arrow now just overrides the file
  size, but both need to be shown one after the other

* Correctly show symlink arrow and file size next

This commit includes quite some refactoring. The shortening of the
left and right part of an entry in a dir list and the shortening of the
file name are separated into functions which don't directly write to the
buffer but just return strings. That way, they get testable and further
enhancements like different line modes should be easier to implement.

* fix review findings

* better file name truncation

File names are only truncated at grapheme border, while the width is
calculated on the actual nuber of terminal cells.

* more adequate test organization

* more code structure