Fix: DirList::need_update considers mtime change (#110)
* Fix: DirList::need_update considers mtime change

Restores behavior which has been removed in commit 7d8b6dd.

`JoshutoDirList::need_update` does not only consider explicit
deprecation but also a changed mtime from the file system.

* Create_or_soft_update for parent on tab switch

When switching a tab, create_or_soft_update is also applied on the
parent directory (shown in the left most pane). Doing so updates the
left most pane when appearing after a tab-switch in case the content
has changed.

* Do soft-reload when opening directory

By issuing a soft-reload when opening a directory with the "open"
command, changed, deleted or added files/dirs cause an update of the
dir-lists of all three panes.