Update view on filesystem changes (#112)
* Update view on filesystem changes

AppContext owns an INotifyWatcher, which watches the visible directories
(max. three). The list of directories to watch is updated on every
event. If a filesystem change is announced, a "soft-update" of the
current view is issued.

* Specific "watchers" for Linux and MacOS

For supervision of filesystem changes, specific `notify::Watcher`
implementations are used for Linux and MacOS.

Builds will likely fail for other OSes as
`notify::recommended_watcher` may return some specific watcher
implementation but the code just expects a `NullWatcher`.

* remove dead code