Add script hooks for image previews
This adds two configuration options, `preview_shown_hook_script`
and `preview_removed_hook_script` in the preview-section where the
user can configure external scripts which are called when a file preview
is shown or when the file preview widget disappears, respectively.

The `preview_shown_hook_script` gets the path of the file being previwed
as first argument and then the x and y coordinate and the width an
height of the preview area as second to fifth parameters.

To get the geometry information back from the UI rendering process, a
new struct `RenderResult` has been introduced which is given as mutable
reference to the rendering function.

The main-loop then checks for a changed preview and calls the hook
scripts.

Additionally, documentation has been added how these hooks can be used
to implement image previews.