Continuous scrolling  (#118)
* Continuous scrolling

The scrolling behavior is changed from “paging” to a continuous
scrolling. Joshuto keeps a buffer from the cursor to each end of the
list, which is configured by `[display] scroll_offset`. If the terminal
height is too small to keep the distance, the buffer is set to a value
that assures that the cursor is at least as close to the end the user is
scrolling towards as to the other end of the visible list.

If the window is resized and the cursor jumps out of scope, the viewport
is adjusted when changing the index next time.

Possible improvements:
* Issue a viewport update on terminal geometry change
* When scrolling down to the bottom, don't allow an empty section
  beneath the last entry

* Update documentation for scroll_offset

* remove unused variable

* keep viewport index when replacing dirlist

* Don't keep copy of scroll_offset in JoshutoDirList

* sanity: remove obsolete parameter