Add linemode feature and first simple linemodes
This adds the basic ability to change the representation of files in the
TUI detailed dir list by letting the user choose a certain "linemode".

As of now, there are only three simple linemodes, all showing the
filename on the left side and a preceding symlink indicator in the
beginning of the right label as usual.

The "size" linemode shows the right label as it has been up to now.
The "mtime" linemode shows the mtime of the file or directory.
The "sizemtime" linemodes combines both meta-data.

The user can change the linemode like so:
```
:linemode [size|mtime|sizemtime]
```

Default keybindings have been added:
``ms`` for *size*, ``mm`` for *mtime*, and ``mM`` for *sizemtime*.

The selected linemode is specific for each tab.

A new tab always starts with the *size* linemode.

Possible enhancements:
* Move the code for factoring a label out of the
  ``tui_dirlist_detailed`` module to some UI-independent module
* Add a configuration option for the default linemode for new tabs
* Add further simple linemodes
* Generic support for linemodes with only one “full line label”
* Add support for custom linemodes where a label is constructed by an
  external script