new_tab can open specific directories
The `new_tab` command got enhanced and can now open a user-specified
directory, or the directory of the current tab, or the directory
currently highlighted by the cursor.

* `new_tab some-dir` opens new tab with directory `some-dir`
* `new_tab --current` opens new tab with the same directory as the current tab
* `new_tab --cursor` opens new tab with the directory which is currently marked by the cursor

The `keymap.toml` documentation got adapted accordingly.

The default key-map also got enhanced/changed:

* `ctrl-t`: opens tab in default directory
* `shift-t`: opens tab in directory of current tab
* `alt-t`: opens tab in directory marked by cursor

Like before, all new tabs get activated when created. Opening tabs “in
the background” is a possible enhancement for the future and might be
especially useful for the `--cursor` variant.