Tab theme configurable
Text style for tabs can be configured within
`theme.toml` with the dotted keys `tab.active` and `tab.inactive`.

The default style is the same as the former hard-coded style.

`theme.toml` has been restructured and comments have been
added to give a short documentation on each style item.

The documentation (`theme.toml.md`) has been enhanced, and
the excerpt of `theme.toml` with additional comments has
been removed from there.

That way it should be easier for devs to keep docs and the default
`theme.toml` in sync, and easier for users when the detail comments
can be found in the configuration file itself.

The conversion from `AppStyle` to `tui::style::Style` is now provided as
a function of `AppStyle` and a `From`-implementation for `Style` to
avoid having the same factory code repeating all over.